Significant Changes to BaseApp - In order for Kivy to play nice with threads and queues in a relibable way - it needed to be the base thread - Otherwise it seems to randomly crash due to its use of OpenGL for drawing interface items
This has some MAJOR consequences
- All threads are created during initialization of Kivy GUI App and are part of it
- Threaded actions for the GUI itself need to be done with the kivy.clock object
- Only one window is possible
- Pygame needs to get unloaded first (This may conflict with Communications threads - not sure)

From my testing it seems that all of the threads work properly and the mailboxes/queues managed by the communication thread works properly.

PS: In order to use Kivy video properly I had to repackage it. Please download it and follow instructions on the USST Google Drive under design\software.

I am open to discussion on choosing a different GUI toolkit but it seems like all the major ones need to be a base thread.